requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0
